First Project Machine Learning FakesNews
This project uses a machine learning model to detect fake news. The fake news, or fass information, may cause a significant impact on the society and promote trompeuses of information. This project uses a machine learning model to detect fake news. The fake news, or fass information, may cause a significant impact on the society and promote trompeuses of information.